Refactor user management pages: consolidate user addition and organization details, remove redundant files, and enhance user role editing functionality.
